ai-service/requirements.txt
# ===== Core API =====
fastapi==0.110.0
uvicorn==0.27.1

# ===== Embedding (CPU-only) =====
sentence-transformers==2.5.1
transformers==4.36.2
torch>=2.2.0

# ===== Vector Search =====
faiss-cpu>=1.8.0

# ===== Minimal deps =====
numpy>=1.26.4
scikit-learn==1.4.0

# ===== HuggingFace =====
datasets>=2.16.0
huggingface-hub>=0.20.0
